basic quantile mapping works along with test_qm

+++ src/objects/data_structures.f90
@@ -69,7 +69,7 @@ module data_structures
     ! type to store quantile mapping data
     ! ------------------------------------------------
     type qm_correction_type
-        integer, allocatable, dimension(:) :: start_idx, end_idx
+        real, allocatable, dimension(:) :: start_idx, end_idx
         real, allocatable, dimension(:) :: slope, offset
     end type qm_correction_type
